Improve Bangcle Check
The so files generated by Bangcle are now mainly `libDexHelper.so`, `libDexHelper-x86.so`, `libdexjni.so`, and the so files with the same name added by other applications are mostly `libDexHelper.so`. Other so files generated by Bangcle in the rules of LibChecker may be the other version of Bangcle. The apk entry class of Bangcle in other rules is `com.secshell.shellwrapper.SecAppWrapper` according to the document online, which is not consistent with the current, so this can prevent the other version of Bangcle from being affected.