Add a tracker for applist large diff (#903)
In order to better fix the problem of large differences in app list, we have introduced a new Exception to track the problem. We only collect the total number of your apps, as well as the total number of new and deleted apps exported internally, and no other information will be collected.

This tracker will be removed when a new version is released.